+++ python/rmm/tests/test_rmm.py
@@ -250,6 +250,18 @@ def test_rmm_device_buffer_pickle_roundtrip(hb):
     db2 = pickle.loads(pb)
     hb2 = db2.tobytes()
     assert hb == hb2
+    # out-of-band
+    if pickle.HIGHEST_PROTOCOL >= 5:
+        db = rmm.DeviceBuffer.to_device(hb)
+        buffers = []
+        pb2 = pickle.dumps(db, protocol=5, buffer_callback=buffers.append)
+        del db
+        assert len(buffers) == 1
+        assert isinstance(buffers[0], pickle.PickleBuffer)
+        assert bytes(buffers[0]) == hb
+        db3 = pickle.loads(pb2, buffers=buffers)
+        hb3 = db3.tobytes()
+        assert hb3 == hb
 
 
 def test_rmm_cupy_allocator():